viewpage implemented for category
