+++ app/src/main/java/com/example/android/miwok/CategoryAdapter.java
@@ -1,0 +1,56 @@
+package com.example.android.miwok;
+
+import android.content.Context;
+import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentManager;
+import android.support.v4.app.FragmentPagerAdapter;
+
+public class CategoryAdapter extends FragmentPagerAdapter {
+    /** Context of the app */
+    private Context mContext;
+
+    public CategoryAdapter(Context context, FragmentManager fm) {
+        super(fm);
+        mContext = context;
+    }
+
+    @Override
+    public Fragment getItem(int position) {
+        if (position == 0) {
+            return new NumberFragment();
+        } else if (position == 1){
+            return new ColorFragment();
+        } else if (position == 2) {
+            return new FamilyFragment();
+        } else {
+            return new PhrasesFragment();
+        }
+    }
+
+    /**
+     * This method may be called by the ViewPager to obtain a title string
+     * to describe the specified page. This method may return null
+     * indicating no title for this page. The default implementation returns
+     * null.
+     *
+     * @param position The position of the title requested
+     * @return A title for the requested page
+     */
+    @Override
+    public CharSequence getPageTitle(int position) {
+        if (position == 0) {
+            return mContext.getString(R.string.category_numbers);
+        } else if (position == 1) {
+            return mContext.getString(R.string.category_family);
+        } else if (position == 2) {
+            return mContext.getString(R.string.category_colors);
+        } else {
+            return mContext.getString(R.string.category_phrases);
+        }
+    }
+
+    @Override
+    public int getCount() {
+        return 4;
+    }
+}
