+++ app/src/main/java/com/example/android/miwok/ColorsActivity.java
@@ -12,109 +12,19 @@ import android.widget.ListView;
 import java.util.ArrayList;
 
 public class ColorsActivity extends AppCompatActivity {
-    private MediaPlayer mMediaPlayer;
-    private AudioManager mAudioManager;
-    private MediaPlayer.OnCompletionListener mCompleteListener = new MediaPlayer.OnCompletionListener() {
-        @Override
-        public void onCompletion(MediaPlayer mediaPlayer) {
-            ColorsActivity.this.releaseMediaPlayer();
-        }
-    };
-
-    private AudioManager.OnAudioFocusChangeListener mOnAudioFocusChangeListener = new AudioManager.OnAudioFocusChangeListener() {
-        @Override
-        public void onAudioFocusChange(int focusChange) {
-            if (focusChange == AudioManager.AUDIOFOCUS_LOSS_TRANSIENT || focusChange == AudioManager.AUDIOFOCUS_GAIN_TRANSIENT_MAY_DUCK) {
-                // Pause playback because your Audio Focus was
-                // stolen, but will be back soon.
-                // i.e. for a phone call
-                mMediaPlayer.pause();
-                mMediaPlayer.seekTo(0);
-            } else if (focusChange == AudioManager.AUDIOFOCUS_LOSS) {
-                // Stop playback, because you lost the Audio Focus.
-                // i.e. the user started some other playback app
-                // Remember to unregister your controls/buttons here.
-                // And release the kra — Audio Focus!
-                // You’re done.
-                releaseMediaPlayer();
-            } else if (focusChange == AudioManager.AUDIOFOCUS_GAIN) {
-                // Resume playback, because you hold the Audio Focus
-                // again!
-                // i.e. the phone call ended or the nav directions
-                // are finished
-                // If you implement ducking and lower the volume, be
-                // sure to return it to normal here, as well.
-                mMediaPlayer.start();
-            }
-        }
-    };
-
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.word_list);
-        // Create and setup {@link AudioManager} to request audio focus
-        mAudioManager = (AudioManager) getSystemService(Context.AUDIO_SERVICE);
-        // Create color array
-        final ArrayList<Word> words = new ArrayList<Word>();
-
-        words.add(new Word("red","weṭeṭṭi", R.drawable.color_red, R.raw.color_red));
-        words.add(new Word("green","chokokki", R.drawable.color_green, R.raw.color_green));
-        words.add(new Word("brown","ṭakaakki", R.drawable.color_brown, R.raw.color_brown));
-        words.add(new Word("gray","ṭopoppi", R.drawable.color_gray, R.raw.color_gray));
-        words.add(new Word("black","kululli", R.drawable.color_black, R.raw.color_black));
-        words.add(new Word("white","kelelli", R.drawable.color_white, R.raw.color_white));
-        words.add(new Word("dusty yellow","ṭopiisә", R.drawable.color_dusty_yellow, R.raw.color_dusty_yellow));
-        words.add(new Word("mustard yellow","chiwiiṭә", R.drawable.color_mustard_yellow, R.raw.color_mustard_yellow));
-
-        WordAdapter itemAdapter = new WordAdapter(this, words, R.color.category_colors);
-
-        ListView listview = (ListView) findViewById(R.id.list);
-        listview.setAdapter(itemAdapter);
-
-        listview.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> adapterView, View view, int i, long l) {
-                releaseMediaPlayer();
-                // Request audio focus for playback
-                int result = mAudioManager.requestAudioFocus(mOnAudioFocusChangeListener,
-                        // Use the music stream
-                        AudioManager.STREAM_MUSIC,
-                        // Request permanent Focus
-                        AudioManager.AUDIOFOCUS_GAIN_TRANSIENT);
-                if (result == AudioManager.AUDIOFOCUS_REQUEST_GRANTED) {
-                    mMediaPlayer = MediaPlayer.create(ColorsActivity.this, words.get(i).getAudioResourceId());
-                    mMediaPlayer.start();
-                    mMediaPlayer.setOnCompletionListener(mCompleteListener);
-                }
-            }
-        });
+        setContentView(R.layout.activity_category);
+        getSupportFragmentManager().beginTransaction().replace(
+                R.id.container, new ColorFragment()
+        ).commit();
     }
 
-    /**
-     * Clean up the media player by releasing its resources.
-     */
-    private void releaseMediaPlayer() {
-        // If the media player is not null, then it may be currently playing a sound.
-        if (mMediaPlayer != null) {
-            // Regardless of the current state of the media player, release its resources
-            // because we no longer need it.
-            mMediaPlayer.release();
-
-            // Set the media player back to null. For our code, we've decided that
-            // setting the media player to null is an easy way to tell that the media player
-            // is not configured to play an audio file at the moment.
-            mMediaPlayer = null;
-            // Regardless of whether or not we were granted audio focus, abandon it. This also
-            // unregister the audio focus change listener , so we don't get any more callback.
-            mAudioManager.abandonAudioFocus(mOnAudioFocusChangeListener);
-        }
-    }
 
     @Override
     protected void onStop() {
         super.onStop();
-        releaseMediaPlayer();
     }
 }
