+++ app/src/main/java/com/example/android/miwok/FamilyActivity.java
@@ -12,110 +12,18 @@ import android.widget.ListView;
 import java.util.ArrayList;
 
 public class FamilyActivity extends AppCompatActivity {
-    private MediaPlayer mMediaPlayer;
-    private AudioManager mAudioManager;
-    private MediaPlayer.OnCompletionListener mCompleteListener = new MediaPlayer.OnCompletionListener() {
-        @Override
-        public void onCompletion(MediaPlayer mediaPlayer) {
-            FamilyActivity.this.releaseMediaPlayer();
-        }
-    };
-
-    private AudioManager.OnAudioFocusChangeListener mOnAudioFocusChangeListener = new AudioManager.OnAudioFocusChangeListener() {
-        @Override
-        public void onAudioFocusChange(int focusChange) {
-            if (focusChange == AudioManager.AUDIOFOCUS_LOSS_TRANSIENT || focusChange == AudioManager.AUDIOFOCUS_GAIN_TRANSIENT_MAY_DUCK) {
-                // Pause playback because your Audio Focus was
-                // stolen, but will be back soon.
-                // i.e. for a phone call
-                mMediaPlayer.pause();
-                mMediaPlayer.seekTo(0);
-            } else if (focusChange == AudioManager.AUDIOFOCUS_LOSS) {
-                // Stop playback, because you lost the Audio Focus.
-                // i.e. the user started some other playback app
-                // Remember to unregister your controls/buttons here.
-                // And release the kra — Audio Focus!
-                // You’re done.
-                releaseMediaPlayer();
-            } else if (focusChange == AudioManager.AUDIOFOCUS_GAIN) {
-                // Resume playback, because you hold the Audio Focus
-                // again!
-                // i.e. the phone call ended or the nav directions
-                // are finished
-                // If you implement ducking and lower the volume, be
-                // sure to return it to normal here, as well.
-                mMediaPlayer.start();
-            }
-        }
-    };
-
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.word_list);
-        // Create and setup {@link AudioManager} to request audio focus
-        mAudioManager = (AudioManager) getSystemService(Context.AUDIO_SERVICE);
-        // Create color ArrayList
-        final ArrayList<Word> words = new ArrayList<Word>();
+        setContentView(R.layout.activity_category);
+        getSupportFragmentManager().beginTransaction().replace(
+                R.id.container, new FamilyFragment()
+        ).commit();
 
-        words.add(new Word("father","әpә", R.drawable.family_father, R.raw.family_father));
-        words.add(new Word("mother","әṭa", R.drawable.family_mother, R.raw.family_mother));
-        words.add(new Word("son","angsi", R.drawable.family_son, R.raw.family_son));
-        words.add(new Word("daughter","tune", R.drawable.family_daughter, R.raw.family_daughter));
-        words.add(new Word("older brother","taachi", R.drawable.family_older_brother, R.raw.family_older_brother));
-        words.add(new Word("younger brother","chalitti", R.drawable.family_younger_brother, R.raw.family_younger_brother));
-        words.add(new Word("older sister","teṭe", R.drawable.family_older_sister, R.raw.family_older_sister));
-        words.add(new Word("younger sister","kolliti", R.drawable.family_younger_sister, R.raw.family_younger_sister));
-        words.add(new Word("grandmother ","ama", R.drawable.family_grandmother, R.raw.family_grandmother));
-
-        WordAdapter itemAdapter = new WordAdapter(this, words, R.color.category_family);
-
-        ListView listview = (ListView) findViewById(R.id.list);
-        listview.setAdapter(itemAdapter);
-
-        listview.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> adapterView, View view, int i, long l) {
-                releaseMediaPlayer();
-                // Request audio focus for playback
-                int result = mAudioManager.requestAudioFocus(mOnAudioFocusChangeListener,
-                        // Use the music stream
-                        AudioManager.STREAM_MUSIC,
-                        // Request permanent Focus
-                        AudioManager.AUDIOFOCUS_GAIN_TRANSIENT);
-                if (result == AudioManager.AUDIOFOCUS_REQUEST_GRANTED) {
-                    mMediaPlayer = MediaPlayer.create(FamilyActivity.this, words.get(i).getAudioResourceId());
-                    mMediaPlayer.start();
-                    mMediaPlayer.setOnCompletionListener(mCompleteListener);
-                }
-            }
-        });
     }
-
-    /**
-     * Clean up the media player by releasing its resources.
-     */
-    private void releaseMediaPlayer() {
-        // If the media player is not null, then it may be currently playing a sound.
-        if (mMediaPlayer != null) {
-            // Regardless of the current state of the media player, release its resources
-            // because we no longer need it.
-            mMediaPlayer.release();
-
-            // Set the media player back to null. For our code, we've decided that
-            // setting the media player to null is an easy way to tell that the media player
-            // is not configured to play an audio file at the moment.
-            mMediaPlayer = null;
-            // Regardless of whether or not we were granted audio focus, abandon it. This also
-            // unregister the audio focus change listener , so we don't get any more callback.
-            mAudioManager.abandonAudioFocus(mOnAudioFocusChangeListener);
-        }
-    }
-
     @Override
     protected void onStop() {
         super.onStop();
-        releaseMediaPlayer();
     }
 }
