+++ app/src/main/res/values/strings.xml
@@ -26,5 +26,7 @@
     <string name="category_numbers">Numbers</string>
 
     <!-- Category name for the vocabulary words for family members [CHAR LIMIT=20] -->
-    <string name="category_family">Family Members</string>
+    <string name="category_family">Family</string>
+    <!-- TODO: Remove or change this placeholder text -->
+
 </resources>
